+++ GameOfUr/src/main/java/model/Dice.java
@@ -18,7 +18,7 @@ public final class Dice
     /**
     * A helper list that contains the weighted sum of each side.
     */
-    private float[] weightedSumList;
+    private int[] weightedSumList;
     /**
     * The amount of sides in current dice.
     */
@@ -29,7 +29,7 @@ public final class Dice
      * @param sides Amount of sides in dice.
      * @param probabilities An array with the probability of each side.
      */
-    public Dice(int sides, float[] probabilities) {
+    public Dice(int sides, int[] probabilities) {
         sideAmount = sides;
         createHelperList(probabilities);
     }
@@ -37,8 +37,8 @@ public final class Dice
      * Creates a helper list with the weighted sum of each side.
      * @param probabilities An array with the probability of each side.
      */
-    private void createHelperList(float[] probabilities) {
-        weightedSumList = new float[sideAmount];
+    private void createHelperList(int[] probabilities) {
+        weightedSumList = new int[sideAmount];
         weightedSumList[0] = probabilities[0];
         for (int sideIndex = 1; sideIndex < sideAmount; sideIndex++) {
             weightedSumList[sideIndex] =
@@ -47,7 +47,7 @@ public final class Dice
     }
     /**
      * Throws the dice with given probabilities.
-     * Based on http://censore.blogspot.com/2014/08/simulate-roll-of-bias-or-weighted-dice.HTML
+     * Adapted from http://censore.blogspot.com/2014/08/simulate-roll-of-bias-or-weighted-dice.HTML
      * @return random number.
      */
     public int throwDice() {
